+++ server/Common/ConnectMgr.cpp
@@ -4,7 +4,6 @@
 #include "msgbase.h"
 #include "GlobalDefine.h"
 
-#include "MainType.h"
 #include "ServerType.h"
 #include "ServerMsg.pb.h"
 
@@ -142,7 +141,7 @@ const char *CConnectMgr::GetMsgNumInfo()
 	return tempbuf;
 }
 
-bool CConnectMgr::SendMsgToServer(int nServerID, google::protobuf::Message &pMsg, int maintype, int subtype, int32 nClientID)
+bool CConnectMgr::SendMsgToServer(int nServerID, google::protobuf::Message &pMsg, int maintype, int subtype, int64 nClientID)
 {
 	connector *con = FindConnect(nServerID);
 	if (con)
@@ -152,7 +151,7 @@ bool CConnectMgr::SendMsgToServer(int nServerID, google::protobuf::Message &pMsg
 	return false;
 }
 
-bool CConnectMgr::SendMsgToServer(int nServerID, Msg &pMsg, int32 nClientID)
+bool CConnectMgr::SendMsgToServer(int nServerID, Msg &pMsg, int64 nClientID)
 {
 	connector *con = FindConnect(nServerID);
 	if (con)
@@ -162,7 +161,7 @@ bool CConnectMgr::SendMsgToServer(int nServerID, Msg &pMsg, int32 nClientID)
 	return false;
 }
 
-bool CConnectMgr::SendMsgToServer(connector *con, google::protobuf::Message &pMsg, int maintype, int subtype, int32 clientid)
+bool CConnectMgr::SendMsgToServer(connector *con, google::protobuf::Message &pMsg, int maintype, int subtype, int64 clientid)
 {
 	assert(con);
 	assert(clientid >= 0);
@@ -174,7 +173,7 @@ bool CConnectMgr::SendMsgToServer(connector *con, google::protobuf::Message &pMs
 	return SendMsg(con, pMsg, maintype, subtype, &tail, sizeof(tail));
 }
 
-bool CConnectMgr::SendMsgToServer(connector *con, Msg &pMsg, int32 clientid)
+bool CConnectMgr::SendMsgToServer(connector *con, Msg &pMsg, int64 clientid)
 {
 	assert(con);
 	assert(clientid >= 0);
