+++ server/GameServer/GameGatewayMgr.h
@@ -23,8 +23,8 @@ public:
 	void ResetMsgNum();
 	const char *GetMsgNumInfo();
 
-	void SendMsgToServer(Msg &pMsg, int nType, int32 nClientID = 0, int nServerID = 0, bool bBroad = false);
-	void SendMsgToServer(google::protobuf::Message &pMsg, int maintype, int subtype, int nType, int32 nClientID = 0, int nServerID = 0, bool bBroad = false);
+	void SendMsgToServer(Msg &pMsg, int nType, int64 nClientID = 0, int nServerID = 0, bool bBroad = false);
+	void SendMsgToServer(google::protobuf::Message &pMsg, int maintype, int subtype, int nType, int64 nClientID = 0, int nServerID = 0, bool bBroad = false);
 
 	void ServerRegisterSucc(int id, int type, const char *ip, int port);
 	void OnConnectDisconnect(serverinfo *info, bool overtime = false);
